requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.1
seaborn>=0.12.2
requests>=2.28.2
beautifulsoup4>=4.12.0
nltk>=3.8.1
textblob>=0.17.1
pyyaml>=6.0
python-telegram-bot>=13.15
newsapi-python>=0.2.7
schedule>=1.2.0
duckdb>=0.9.0
yfinance>=0.2.0
tiktoken>=0.9.0
scipy>=1.13.1
sqlalchemy>=2.0.39
openai>=1.0.0